- Tests and page objects for Breadcrumbs are refactored
