+++ jdi-light-material-ui-tests/src/main/java/io/github/com/pages/navigation/RouterBreadcrumbsPage.java
@@ -8,23 +8,14 @@ import com.epam.jdi.light.material.elements.inputs.Button;
 import java.util.List;
 
 /**
- * To see an example of Accordion web element please visit
- * https://material-ui.com/ru/components/breadcrumbs/
+ * To see an example of a Breadcrumbs web element, please visit
+ * https://material-ui.com/components/breadcrumbs/
  */
 
-public class BreadcrumbsPage extends WebPage {
+public class RouterBreadcrumbsPage extends WebPage {
 
     @UI(".MuiBreadcrumbs-li")
     public static List<UIElement> breadcrumbs;
-
-    @UI(".MuiBreadcrumbs-separator")
-    public static List<UIElement> separators;
-
-    @UI(".MuiBreadcrumbs-li .MuiSvgIcon-root")
-    public static List<UIElement> icons;
-
-    @UI(".MuiSvgIcon-root.jss6")
-    public static Button collapsedButton;
 
     @UI("//span[text()='Trash']")
     public static UIElement trashDisplayBlock;
@@ -34,12 +25,5 @@ public class BreadcrumbsPage extends WebPage {
 
     @UI("//span[text()='Inbox']")
     public static UIElement inboxDisplayBlock;
-
-    @UI("//h3[text()='Material UI']")
-    public static UIElement materialElement;
-
-    @UI("//h3[text()='Core']")
-    public static UIElement coreElement;
-
 
 }
